refactor: 최종 결과를 보여주는 로직을 OutputView로부터 분리하여 ResultView에 구현
